Refactor preferences UI
- Use separate screens for a few main categories
- Categorize preferences
- Reorder preferences/categories/screens
- Rename strings to make more sense
- Add missing descriptions
- Rename "settings" to "preferences"
- Rename "user" to "account"
- Use Normal Capitalization Instead Of Title Casing
- Add icons to preference links and about items
- Remove Utils.getVersionString
