+++ app/src/main/java/org/jellyfin/androidtv/ui/preference/screen/HomePreferencesScreen.kt
@@ -6,7 +6,7 @@ import org.jellyfin.androidtv.preference.PreferenceStore
 import org.jellyfin.androidtv.preference.UserSettingPreferences
 import org.jellyfin.androidtv.ui.preference.dsl.OptionsFragment
 import org.jellyfin.androidtv.ui.preference.dsl.enum
-import org.jellyfin.androidtv.ui.preference.dsl.optionsScreen
+import org.jellyfin.androidtv.ui.preference.dsl.lazyOptionsScreen
 import org.koin.android.ext.android.inject
 
 class HomePreferencesScreen : OptionsFragment() {
@@ -15,27 +15,26 @@ class HomePreferencesScreen : OptionsFragment() {
 	override val stores: Array<PreferenceStore>
 		get() = arrayOf(userSettingPreferences)
 
-	override val screen
-		get() = optionsScreen {
-			setTitle(R.string.home_prefs)
+	override val screen by lazyOptionsScreen {
+		setTitle(R.string.home_prefs)
 
-			category {
-				title = getString(R.string.home_sections)
+		category {
+			setTitle(R.string.home_sections)
 
-				arrayOf(
-					UserSettingPreferences.homesection0,
-					UserSettingPreferences.homesection1,
-					UserSettingPreferences.homesection2,
-					UserSettingPreferences.homesection3,
-					UserSettingPreferences.homesection4,
-					UserSettingPreferences.homesection5,
-					UserSettingPreferences.homesection6,
-				).forEachIndexed { index, section ->
-					enum<HomeSectionType> {
-						title = getString(R.string.home_section_i, index + 1)
-						bind(userSettingPreferences, section)
-					}
+			arrayOf(
+				UserSettingPreferences.homesection0,
+				UserSettingPreferences.homesection1,
+				UserSettingPreferences.homesection2,
+				UserSettingPreferences.homesection3,
+				UserSettingPreferences.homesection4,
+				UserSettingPreferences.homesection5,
+				UserSettingPreferences.homesection6,
+			).forEachIndexed { index, section ->
+				enum<HomeSectionType> {
+					title = getString(R.string.home_section_i, index + 1)
+					bind(userSettingPreferences, section)
 				}
 			}
 		}
+	}
 }
